auth

+++ app/Http/Controllers/TaskController.php
@@ -151,6 +151,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('delete-task');
+        
         $task->delete();
 
         return redirect()->route('task.index')->with('success', 'Task deleted successfully.');
